# Project Files/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * House Rent API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ListingType = typeof ListingType[keyof typeof ListingType];


export const ListingType = {
  apartment: 'apartment',
  house: 'house',
  condo: 'condo',
  studio: 'studio',
  townhouse: 'townhouse',
} as const;

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  available: 'available',
  rented: 'rented',
} as const;

export interface Listing {
  id: number;
  title: string;
  description: string;
  address: string;
  city: string;
  state: string;
  price: number;
  bedrooms: number;
  bathrooms: number;
  area: number;
  type: ListingType;
  status: ListingStatus;
  featured: boolean;
  amenities: string[];
  images: string[];
  createdAt: string;
}

export type ListingInputType = typeof ListingInputType[keyof typeof ListingInputType];


export const ListingInputType = {
  apartment: 'apartment',
  house: 'house',
  condo: 'condo',
  studio: 'studio',
  townhouse: 'townhouse',
} as const;

export type ListingInputStatus = typeof ListingInputStatus[keyof typeof ListingInputStatus];


export const ListingInputStatus = {
  available: 'available',
  rented: 'rented',
} as const;

export interface ListingInput {
  /** @minLength 1 */
  title: string;
  description: string;
  address: string;
  city: string;
  state: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  bedrooms: number;
  /** @minimum 0 */
  bathrooms: number;
  /** @minimum 0 */
  area: number;
  type: ListingInputType;
  status?: ListingInputStatus;
  featured?: boolean;
  amenities?: string[];
  images?: string[];
}

export type ListingUpdateType = typeof ListingUpdateType[keyof typeof ListingUpdateType];


export const ListingUpdateType = {
  apartment: 'apartment',
  house: 'house',
  condo: 'condo',
  studio: 'studio',
  townhouse: 'townhouse',
} as const;

export type ListingUpdateStatus = typeof ListingUpdateStatus[keyof typeof ListingUpdateStatus];


export const ListingUpdateStatus = {
  available: 'available',
  rented: 'rented',
} as const;

export interface ListingUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  address?: string;
  city?: string;
  state?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  bedrooms?: number;
  /** @minimum 0 */
  bathrooms?: number;
  /** @minimum 0 */
  area?: number;
  type?: ListingUpdateType;
  status?: ListingUpdateStatus;
  featured?: boolean;
  amenities?: string[];
  images?: string[];
}

export interface CityCount {
  city: string;
  count: number;
}

export interface ListingsSummary {
  total: number;
  available: number;
  rented: number;
  avgPrice: number;
  totalInquiries: number;
  cityCounts: CityCount[];
}

export interface Inquiry {
  id: number;
  listingId: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
  createdAt: string;
}

export interface InquiryInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  /** @minLength 1 */
  message: string;
}

export type GetListingsParams = {
city?: string;
minPrice?: number;
maxPrice?: number;
bedrooms?: number;
type?: GetListingsType;
status?: GetListingsStatus;
search?: string;
};

export type GetListingsType = typeof GetListingsType[keyof typeof GetListingsType];


export const GetListingsType = {
  apartment: 'apartment',
  house: 'house',
  condo: 'condo',
  studio: 'studio',
  townhouse: 'townhouse',
} as const;

export type GetListingsStatus = typeof GetListingsStatus[keyof typeof GetListingsStatus];


export const GetListingsStatus = {
  available: 'available',
  rented: 'rented',
} as const;