# Project Files/lib/api-zod/src/generated/types/cityCount.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * House Rent API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CityCount {
  city: string;
  count: number;
}