# Project Files/lib/api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * House Rent API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Inquiry {
  id: number;
  listingId: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
  createdAt: Date;
}