# Project Files/lib/api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * House Rent API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InquiryInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  /** @minLength 1 */
  message: string;
}