# Project Files/lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * House Rent API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ListingStatus } from './listingStatus';
import type { ListingType } from './listingType';

export interface Listing {
  id: number;
  title: string;
  description: string;
  address: string;
  city: string;
  state: string;
  price: number;
  bedrooms: number;
  bathrooms: number;
  area: number;
  type: ListingType;
  status: ListingStatus;
  featured: boolean;
  amenities: string[];
  images: string[];
  createdAt: Date;
}