# Project Files/lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * House Rent API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ListingInputStatus } from './listingInputStatus';
import type { ListingInputType } from './listingInputType';

export interface ListingInput {
  /** @minLength 1 */
  title: string;
  description: string;
  address: string;
  city: string;
  state: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  bedrooms: number;
  /** @minimum 0 */
  bathrooms: number;
  /** @minimum 0 */
  area: number;
  type: ListingInputType;
  status?: ListingInputStatus;
  featured?: boolean;
  amenities?: string[];
  images?: string[];
}