# Project Files/lib/api-zod/src/generated/types/listingsSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * House Rent API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CityCount } from './cityCount';

export interface ListingsSummary {
  total: number;
  available: number;
  rented: number;
  avgPrice: number;
  totalInquiries: number;
  cityCounts: CityCount[];
}